% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{addNLTKStopwords}
\alias{addNLTKStopwords}
\title{add NLTK stopwords to the current list of stopwords}
\usage{
addNLTKStopwords(nltk_path, stop.words)
}
\arguments{
\item{nltk_path}{: path to nltk stopwords file}

\item{stopwords}{: stopwords list}
}
\value{
a vector of stopwords
}
\description{
add NLTK stopwords to the current list of stopwords
}

