% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{centerScaleData}
\alias{centerScaleData}
\title{center and scale a numeric dataset}
\usage{
centerScaleData(df_numeric)
}
\arguments{
\item{df_numeric}{: a dataframe with only numeric columns}
}
\value{
scaled dataframe
}
\description{
center and scale a numeric dataset
}

