% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{cleanTitleString}
\alias{cleanTitleString}
\title{Get a clean string which is based on hueristic for document titles}
\usage{
cleanTitleString(string)
}
\arguments{
\item{string}{: title string}
}
\value{
string
}
\description{
Get a clean string which is based on hueristic for document titles
}

