% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{createFreqOfKeywordsTerms}
\alias{createFreqOfKeywordsTerms}
\title{Get freq of keywords in dataset}
\usage{
createFreqOfKeywordsTerms(input_path, output_path)
}
\arguments{
\item{input_path}{: path to csv file}

\item{output_path}{: path where the  output csv file is to be written}
}
\value{
integer value
}
\description{
Get freq of keywords in dataset
}

