% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{createTextVectorFromDataset}
\alias{createTextVectorFromDataset}
\title{createTextVectorFromDataset from title}
\usage{
createTextVectorFromDataset(df, removePunctuation = FALSE,
  withTextColumn = FALSE)
}
\arguments{
\item{removePunctuation}{: The default is FALSE. If true will remove all the punctuations. All entities other than space, numbers and characters}

\item{withTextColumn}{: Whether the input dataframe has text column. The default is FALSE}

\item{df:}{A dataframe with column "title" or optional "text" column}
}
\value{
vector of clean text for each row in dataframe
}
\description{
Get a clean text vector from the columns "title" without punctuations
}

