% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getAcronymRatio}
\alias{getAcronymRatio}
\title{get Acronym Ratio}
\usage{
getAcronymRatio(string)
}
\arguments{
\item{string}{: Get input string}
}
\value{
decimal value
}
\description{
get Acronym Ratio
}

