% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{getAUCandPlotROC}
\alias{getAUCandPlotROC}
\title{Generic method to quickly plot ROC curve and print the AUC value}
\usage{
getAUCandPlotROC(predictedValues, trueValues)
}
\arguments{
\item{predictedValues:}{These is the vector of predicted value}

\item{trueValues:}{These are true outcomes against the predicted value}
}
\value{
None
}
\description{
Generic method to quickly plot ROC curve and print the AUC value
}

