% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getBOWfeatures_Binary}
\alias{getBOWfeatures_Binary}
\title{get BOW features binary - y/n}
\usage{
getBOWfeatures_Binary(dataset, low_freq_thresh, high_freq_thresh)
}
\arguments{
\item{dataset}{: input dataframe}
}
\value{
dataframe
}
\description{
get BOW features binary - y/n
}

