% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getBOWfeatures_freq}
\alias{getBOWfeatures_freq}
\title{get bag of words features with frew}
\usage{
getBOWfeatures_freq(dataset, test_dataset)
}
\arguments{
\item{dataset}{: dataframe}
}
\value{
dataframe .
}
\description{
get bag of words features with frew
}

