% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getBOWfeaturesTestDataset}
\alias{getBOWfeaturesTestDataset}
\title{get bag of words test dataset}
\usage{
getBOWfeaturesTestDataset(dataset, keywords)
}
\arguments{
\item{dataset:}{dataframe}

\item{keywords:}{the keywords}
}
\value{
dataframe
}
\description{
get bag of words test dataset
}

