% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getBOWKeywords}
\alias{getBOWKeywords}
\title{get bagofwords keywords from a dataset}
\usage{
getBOWKeywords(dataset)
}
\arguments{
\item{dataset:}{dataframe with title and id column}
}
\value{
keywords
}
\description{
get bagofwords keywords from a dataset
}

