% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getCamelCaseKeywords}
\alias{getCamelCaseKeywords}
\title{Get camel case keywords from string}
\usage{
getCamelCaseKeywords(string)
}
\arguments{
\item{string:}{input string}
}
\value{
a vector of camel case string words
}
\description{
Get camel case keywords from string
}

