% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getCapLettersToCharactersRatio}
\alias{getCapLettersToCharactersRatio}
\title{get capital letters to character ratio}
\usage{
getCapLettersToCharactersRatio(string)
}
\arguments{
\item{string}{: input string}
}
\value{
decimal value
}
\description{
get capital letters to character ratio
}

