% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getCharacterVector}
\alias{getCharacterVector}
\title{Get Character Vector}
\usage{
getCharacterVector(string)
}
\arguments{
\item{string}{: this is the text which needs to be transformed to character vector}
}
\value{
character vector
}
\description{
Get words from string
}

