% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getCorpusFromTextVector}
\alias{getCorpusFromTextVector}
\title{get corpus from text vector}
\usage{
getCorpusFromTextVector(textVector)
}
\arguments{
\item{textVector:}{a vector of text documents from which to create the corpus}
}
\value{
a corpus entity of tm package
}
\description{
get corpus from text vector
}

