% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getCountByPattern}
\alias{getCountByPattern}
\title{get count by pattern}
\usage{
getCountByPattern(string, pattern, splitby = "\\\\s")
}
\arguments{
\item{string:}{input string}

\item{pattern:}{the regex pattern to apply on each word of the string}

\item{splitby:}{instead of splitting the string into words, it can be split using this regex pattern}
}
\value{
an integer value
}
\description{
get count by pattern
}

