% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getDataframeFromWordVector}
\alias{getDataframeFromWordVector}
\title{Creates a dataframe whose columns are words in wordvector where the rows are all set to 0.}
\usage{
getDataframeFromWordVector(wordVec, nrows = 1)
}
\arguments{
\item{nrows}{: number of rows. By default is 1}

\item{wordvec}{: a word vector}
}
\value{
a datframe
}
\description{
Creates a dataframe whose columns are words in wordvector where the rows are all set to 0.
}

