% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getDataSetWithFeatures}
\alias{getDataSetWithFeatures}
\title{get data set with features}
\usage{
getDataSetWithFeatures(dataset, funToGetBOWFeatures, test_dataset,
  funargs = NULL, otherargs = NULL)
}
\arguments{
\item{dataset}{: dataframe which would be the train dataset}

\item{funToGetBOWFeatures}{: function to get bag of words features from dataframe}

\item{test_dataset}{: dataframe which would be the test dataset}

\item{stopwords}{: Stop word}
}
\value{
a list of dataframes with various bag of words features. First one is the test and second is train dataframe
}
\description{
get data set with features
}

