% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_topicmodels.R
\name{getDocProbabilityDistribution}
\alias{getDocProbabilityDistribution}
\title{Get the topic probability distribution for the docs}
\usage{
getDocProbabilityDistribution(ldaObj, k)
}
\arguments{
\item{k}{: the number of topics that the algorithm should use to classify documents}
}
\value{
a dataframe
}
\description{
Get the topic probability distribution for the docs
}

