% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{getFeaturesAboveThreshold}
\alias{getFeaturesAboveThreshold}
\title{get the features above a threshold value}
\usage{
getFeaturesAboveThreshold(model, threshold)
}
\arguments{
\item{model}{: model}
}
\description{
get the features above a threshold value
}

