% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getFeaturesForDataset}
\alias{getFeaturesForDataset}
\title{Given a dataset, it returns the dataset updated with features}
\usage{
getFeaturesForDataset(dataset, keywords, stpwords, dbh)
}
\arguments{
\item{dataset}{: Input dataset}

\item{keywords}{: These are the domain keywords unigrams}

\item{stpwords}{: This is the stopwords}

\item{dbh}{: The database handle where tables for feature extraction exists}
}
\description{
The dataset should have a column title. The features are decided based on that
}

