% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getFreqOfKeywordInDataset}
\alias{getFreqOfKeywordInDataset}
\title{Get freq of keywords in dataset}
\usage{
getFreqOfKeywordInDataset(dataset)
}
\arguments{
\item{dataset}{: a dataframe with a "title" column}
}
\value{
integer value
}
\description{
Get freq of keywords in dataset
}

