% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getFreqWordFeaturesFromTdm}
\alias{getFreqWordFeaturesFromTdm}
\title{get freq words from tdm}
\usage{
getFreqWordFeaturesFromTdm(tdm, low_freq_thresh, high_freq_thresh)
}
\arguments{
\item{tdm:}{term document matrix}
}
\value{
vector of words
}
\description{
get freq words from tdm
}

