% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getKeywordWeights}
\alias{getKeywordWeights}
\title{Get the suma of weights of all keywords in string}
\usage{
getKeywordWeights(dbh, string)
}
\arguments{
\item{dbh}{: datbase handle to keywords database}

\item{string}{: input string}
}
\value{
decimal value
}
\description{
Get the suma of weights of all keywords in string
}

