% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_topicmodels.R
\name{getLDATopicForDocs}
\alias{getLDATopicForDocs}
\title{Get LDA top Topic terms for docs}
\usage{
getLDATopicForDocs(ldaObj)
}
\arguments{
\item{ldaObj}{: LDA object returned after LDA run}
}
\description{
Get LDA top Topic terms for docs
}

