% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_topicmodels.R
\name{getLDATopTopicTerms}
\alias{getLDATopTopicTerms}
\title{Get top terms from lda object}
\usage{
getLDATopTopicTerms(ldaObj, count)
}
\arguments{
\item{ldaObj}{: LDA object returned after LDA run}

\item{count}{: The count of top terms to be returned}
}
\value{
vector of top terms
}
\description{
Get top terms from lda object
}

