% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{getModelPerformance}
\alias{getModelPerformance}
\title{show the model performance statistics like AUC, ROC, confusion matrix and more}
\usage{
getModelPerformance(model, trainDataSet, testDataSet)
}
\arguments{
\item{model}{: the model to be evaluated}

\item{trainDataSet}{: the dataset with features used for training the model}

\item{testDataSet:}{the dataset with features used for testing the model}
}
\description{
show the model performance statistics like AUC, ROC, confusion matrix and more
}

