% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getPredictionsForModel}
\alias{getPredictionsForModel}
\title{get predicted results for the model built for text only dataset}
\usage{
getPredictionsForModel(dataset, model, keywords, stpwords, dbh)
}
\arguments{
\item{dataset}{: the dataframe with a title field}

\item{model}{: the model used for making predictions}

\item{keywords}{: the domain keywords used for building features}

\item{stpwords}{: the stopwords list used for building features}

\item{dbh}{: the database handle}
}
\value{
the dataset with prediction as 0 or 1. The values are stored in a pred column
}
\description{
get predicted results for the model built for text only dataset
}

