% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getRatio}
\alias{getRatio}
\title{Apply a regex pattern on each word of a string and find the ratio of matches to word length}
\usage{
getRatio(string, pattern, splitby = "\\\\s")
}
\arguments{
\item{string}{: the input string}

\item{pattern}{: the regex patter to apply on each word of the string}

\item{splitby}{: instead of splitting the string into words, it can be split using this regex pattern}
}
\value{
a decimal value
}
\description{
Apply a regex pattern on each word of a string and find the ratio of matches to word length
}

