% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{getStopWords}
\alias{getStopWords}
\title{Get the stopwords to be used for feature building}
\usage{
getStopWords(nltk_stopwords_path)
}
\arguments{
\item{nltk_stopwords_path}{: Path to the nltk english stopwords file}
}
\value{
stopwords combining the nltk and tm package stopwords
}
\description{
Get the stopwords to be used for feature building
}

