% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getStopWordsCount}
\alias{getStopWordsCount}
\title{Get stop Words Count in string}
\usage{
getStopWordsCount(string, stpwords)
}
\arguments{
\item{string}{: input string}
}
\value{
: count of number of stopwords
}
\description{
: None
}

