% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getStopWordsRatio}
\alias{getStopWordsRatio}
\title{Get the ratio of stop words}
\usage{
getStopWordsRatio(string, stpwords)
}
\arguments{
\item{string}{: input string}

\item{output_path}{: path where the  output csv file is to be written}
}
\value{
integer value
}
\description{
Get the ratio of stop words
}

