% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_stats.R
\name{getSymbolToWordsRatio}
\alias{getSymbolToWordsRatio}
\title{get Symbol to words ratio}
\usage{
getSymbolToWordsRatio(string)
}
\arguments{
\item{string}{: input string}
}
\value{
decimal value
}
\description{
get Symbol to words ratio
}

