% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{getTestTrainDomainDataset}
\alias{getTestTrainDomainDataset}
\title{Get test and train dataset}
\usage{
getTestTrainDomainDataset(path_generic, path_domain)
}
\arguments{
\item{path_generic}{: Path to a generic dataframe with columns title, outcome (0 and 1) and id fields}

\item{path_generic}{: Path to a domain dataframe with columns title, outcome (0 and 1) and id fields}
}
\value{
test and train dataframe
}
\description{
There are two dataset concept
Generic dataset : This is a generic marked dataset with outcome column marked 0 and 1
This is generic so has very few of the terms marked with 1. Around 25% of terms
The negative values in this dataset is of interest since they fit into none of the domain dataset
and can be used the negative case for training
Domain Dataset : This is the domain datset where all the rows are marked 1
In this function we create a new dataset which is combination of the domain dataset and the
negative values from the generic dataset
Then we return the test and train value
}

