% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{getTestTrainGenericDataset}
\alias{getTestTrainGenericDataset}
\title{get the test train generic dataset}
\usage{
getTestTrainGenericDataset(path_generic)
}
\arguments{
\item{path_generic}{: Path to the generic dataset}
}
\value{
a list of values containing the various datasets
}
\description{
get the test train generic dataset
}

