% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getTextVectorAll}
\alias{getTextVectorAll}
\title{Get clean text from dataframe}
\usage{
getTextVectorAll(df)
}
\arguments{
\item{df:}{A dataframe with column "title" and "text"}
}
\value{
vector of clean text for each row in dataframe
}
\description{
Get a clean text vector by merging the columns "title" and "text" of a dataframe
}

