% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getWordFeaturesFromTdm}
\alias{getWordFeaturesFromTdm}
\title{get the feature vector form term document matrix (tdm)}
\usage{
getWordFeaturesFromTdm(tdm)
}
\arguments{
\item{tdm:}{the term document matrix}
}
\value{
a vector of words
}
\description{
get the feature vector form term document matrix (tdm)
}

