% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{getWordFeatureVector}
\alias{getWordFeatureVector}
\title{Apply a function to create feature vector from dataset}
\usage{
getWordFeatureVector(df, fun, arg1 = NULL)
}
\arguments{
\item{df}{: dataframe with a title column}

\item{fun}{: function to apply to create the features from text extracted from dataframe}
}
\value{
a vector of word features extracted by the fun
}
\description{
Apply a function to create feature vector from dataset
}

