% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_transform.R
\name{getWords}
\alias{getWords}
\title{Get Words}
\usage{
getWords(string, splitby = "\\\\s", min = 0)
}
\arguments{
\item{string}{: this is the text which needs to be transformed}

\item{splitby}{: the regex pattern on which the string needs to be split.
The default is split by space like:  "\\s"}

\item{min}{: the min word length}
}
\value{
vector of words
}
\description{
Nothing special
}

