% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_utils.R
\name{isDictWord}
\alias{isDictWord}
\title{Checks if the word is a dict word}
\usage{
isDictWord(dbh, word)
}
\arguments{
\item{dbh:}{sqlite dbh handle. Should have table named words_english}

\item{word:}{the word to check}
}
\value{
TRUE or FALSE
}
\description{
Checks if the word is a dict word
}

