\name{latexrule}
\alias{latexrule}
\title{
  LaTeX Rule.
}
\description{
  Create a LaTeX-rule, including colours.
}
\usage{
latexrule(x, y, col = NULL, x.unit = "cm", y.unit = "cm", noindent = FALSE)
}
\arguments{
  \item{x}{
    numeric
}
  \item{y}{
    numeric
}
  \item{col}{
    character
}
  \item{x.unit}{
    character
}
  \item{y.unit}{
    character
}
  \item{noindent}{
    logical
}
}
\details{
  Experimental. Create LaTeX code that produces rules.
}
\value{
  character
}
\author{
  Enrico Schumann
}
\examples{
## see vignette
}
