% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utils.R
\name{recreateAndRunModelWithSelectedFeatures}
\alias{recreateAndRunModelWithSelectedFeatures}
\title{Given a model, will detect the significant features and rerun the show the model performance statistics
like AUC, ROC, confusion matrix and more}
\usage{
recreateAndRunModelWithSelectedFeatures(model, modelfun, trainDataSet,
  testDataSet)
}
\arguments{
\item{model}{: the model to be evaluated}

\item{modelfun}{: the function to apply to create model}

\item{trainDataSet}{: the dataset with features used for training the model}

\item{testDataSet:}{the dataset with features used for testing the model}
}
\description{
Given a model, will detect the significant features and rerun the show the model performance statistics
like AUC, ROC, confusion matrix and more
}

