% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{removeNonNumericFeatures}
\alias{removeNonNumericFeatures}
\title{remove non numeric features from dataframe}
\usage{
removeNonNumericFeatures(dataset)
}
\description{
remove non numeric features from dataframe
}

