% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_topicmodels.R
\name{runLDAGibbs}
\alias{runLDAGibbs}
\title{Run LDA based on gibbs method}
\usage{
runLDAGibbs(dtm, k = 5)
}
\arguments{
\item{k}{: the number of topics that the algorithm should use to classify documents}
}
\value{
LDA object
}
\description{
Run LDA based on gibbs method
}

