% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{setAllBOWFeaturesInDataFrame}
\alias{setAllBOWFeaturesInDataFrame}
\title{set all the words in a dataframe}
\usage{
setAllBOWFeaturesInDataFrame(df, wordFeatures)
}
\arguments{
\item{df:}{the data frame}

\item{wordFeatures:}{the word features to set in the dataframe column}
}
\value{
updated dataframe
}
\description{
set all the words in a dataframe
}

