% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{setBOWFeatureInDataFrame}
\alias{setBOWFeatureInDataFrame}
\title{set bag of words in a data frame}
\usage{
setBOWFeatureInDataFrame(row, wordFeatures)
}
\arguments{
\item{row}{: the data frame row}

\item{wordFeatures}{: the word features}
}
\value{
the updated row
}
\description{
set bag of words in a data frame
}

