% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{setBOWkeywords}
\alias{setBOWkeywords}
\title{return the dataset with bag of words keywords set}
\usage{
setBOWkeywords(dataset, keywords)
}
\arguments{
\item{dataset}{: dataframe}

\item{keywords}{: keyword}
}
\value{
dataframe
}
\description{
return the dataset with bag of words keywords set
}

