\name{spaces}
\alias{spaces}
\title{
  Create Vectors of White Space
}
\description{
  Create character vectors of white space.
}
\usage{
spaces(n)
}
\arguments{
  \item{n}{
    integer
}
}
\details{

  The function creates a character vector of
  white-space strings. Such vectors are useful, for
  instance, for padding character vectors.

}
\value{
  character
}
\author{
  Enrico Schumann
}
\seealso{
  \code{\link{strexp}}
}
\examples{
spaces(0:3)
}
