\name{textutils-package}
\alias{textutils-package}
\alias{textutils}
\docType{package}
\title{
\packageTitle{textutils}
}
\description{
\packageDescription{textutils}
}
\details{
Functions
}
\author{
\packageAuthor{textutils}

Maintainer: \packageMaintainer{textutils}
}
\keyword{ package }
%% \seealso{

%% }
%% \examples{
%% ~~ simple examples of the most important functions ~~
%% }
