\name{toHTML}
\alias{toHTML}
\title{
  Convert R Objects to HTML
}
\description{
  Convert an R object to an HTML snippet.
}
\usage{
toHTML(x, ...)
}
\arguments{
  \item{x}{
    an object
}
  \item{\dots}{
    arguments passed to methods
}
}
\details{
There exists several toHTML methods.
}
\value{
A character vector.
}
\author{
Enrico Schumann
}
\note{
  Further generics.
}
\seealso{
  \code{\link{toLatex}}
}
\examples{
1+1
}
