\name{textutils-package}
\alias{textutils-package}
\alias{textutils}
\docType{package}
\title{
\packageTitle{textutils}
}
\description{
\packageDescription{textutils}
}
\details{

  The package comprises a number of functions that help with
  manipulating character strings.

  For more information and a complete list of functions, use
      \sQuote{\code{library(help = "textutils")}}.

}
\author{

  \packageAuthor{textutils}

  Maintainer: \packageMaintainer{textutils}

}
\keyword{package}
