\name{toHTML}
\alias{toHTML}
\title{
  Convert R Objects to HTML
}
\description{
  Convert an R object to an HTML snippet.
}
\usage{
toHTML(x, ...)
}
\arguments{
  \item{x}{
    an object
}
  \item{\dots}{
    arguments passed to methods
}
}
\details{

  There exists \code{toHTML} methods in several packages: in
  \pkg{tools}, \pkg{XML}. \pkg{R2HTML} has a \code{HTML} generic.
  
}
\value{
  A character vector.
}
\author{
Enrico Schumann
}
\note{
  Further generics.
}
\seealso{
  \code{\link{toLatex}}
}
%% \examples{
%% ## see returns in package
%% }
